+++ src/ultimate_notion/__init__.py
@@ -49,8 +49,8 @@ from ultimate_notion.file import Emoji, FileInfo
 from ultimate_notion.obj_api.enums import AggFunc, BGColor, CodeLang, Color, NumberFormat, VState
 from ultimate_notion.option import Option, OptionNS
 from ultimate_notion.page import Page
-from ultimate_notion.rich_text import Text, join, math, mention, text
-from ultimate_notion.schema import Property, PropType, Schema, SelfRef
+from ultimate_notion.rich_text import join, math, mention, text
+from ultimate_notion.schema import Property, PropType, Schema, SchemaError, SelfRef
 from ultimate_notion.session import Session
 from ultimate_notion.user import User
 from ultimate_notion.utils import SList
@@ -93,13 +93,13 @@ __all__ = [
     'Quote',
     'SList',
     'Schema',
+    'SchemaError',
     'SelfRef',
     'Session',
     'SyncedBlock',
     'Table',
     'TableOfContents',
     'TableRow',
-    'Text',  # mainly for type hinting
     'ToDoItem',
     'ToggleItem',
     'User',