+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - New: Reading and inserting comments is implemented.
 - Chg: `SList` is part of the public API now, i.e. is in the `ultimate_notion` namespace.
 - Fix: Added missing `link_to_page` subtype `comment_id`.
+- Chg: Remove `Text` as direct import, use `text` instead.
 
 ## Version 0.6, 2024-09-28